globals.h 다시 적용
