+++ include/limit.h
@@ -5,12 +5,5 @@
 ************************************************************************* */
 
 /* Public Procedures */
-long int mana_limit(struct char_data *ch);
-long int hit_limit(struct char_data *ch);
-long int move_limit(struct char_data *ch);
 
-struct title_type {
-	char *title_m;
-	char *title_f;
-	long int exp;
-};
+
