+++ include/spells.h
@@ -1,3 +1,5 @@
+#ifndef _SPELLS_H_
+#define _SPELLS_H_
 /* ************************************************************************
 *  file: spells.h , Implementation of magic spells.       Part of DIKUMUD *
 *  Usage : Spells                                                         *
@@ -206,3 +208,5 @@ struct attack_hit_type {
 	char *singular;
 	char *plural;
 };
+
+#endif /* _SPELLS_H_ */
