+++ lib/motd
@@ -30,5 +30,5 @@
 
 -------------------------------------------------------------------------------
    *** Build Version  : unknown
-   *** Build Time     : 2025-12-18 13:55:09 KST
-   *** Server Started : 2025-12-18 13:55:20 KST
+   *** Build Time     : 2025-12-18 15:24:03 KST
+   *** Server Started : 2025-12-18 15:24:11 KST
