+++ src/act.informative.c
@@ -23,25 +23,6 @@
 
 #include "guild_list.h"		/* by process */
 
-/* extern variables */
-
-extern struct room_data *world;
-extern struct descriptor_data *descriptor_list;
-extern struct char_data *character_list;
-extern struct obj_data *object_list;
-extern char credits[MAX_STRING_LENGTH];
-extern char news[MAX_STRING_LENGTH];
-extern char plan[MAX_STRING_LENGTH];
-extern char wizards[MAX_STRING_LENGTH];		/* cyb */
-extern char *dirs[];
-extern char *where[];
-extern char *color_liquid[];
-extern char *fullness[];
-extern char *spells[];
-
-extern char *guild_names[];
-extern char *connected_types[];
-
 
 /* intern vars */
 char *news_content = NULL; // 뉴스 내용을 저장할 메모리, 251119
@@ -680,7 +661,6 @@ void do_look(struct char_data *ch, char *argument, int cmd)
 		case 8:
 			{
 				if (GET_LEVEL(ch) >= IMO) {
-					extern char *room_bits[];
 					snprintf(buffer, sizeof(buffer), "%s  [%5d] [ ",
 						world[ch->in_room].name,
 						world[ch->in_room].number);
@@ -954,11 +934,8 @@ void do_score(struct char_data *ch, char *argument, int cmd)
 	struct affected_type *aff;
 	struct time_info_data playing_time;
 	char buf[200], buf2[200];
-	struct time_info_data real_time_passed(time_t t2, time_t t1);
 	int tmp;
-	extern byte saving_throws[4][5][IMO + 10];
-	extern struct title_type titles[4][IMO + 4];
-
+	
 	snprintf(buf, sizeof(buf), "You are %d years old.\n\r", GET_AGE(ch));
 	snprintf(buf2, sizeof(buf2), "당신은 %d 살 입니다.\n\r", GET_AGE(ch));
 	send_to_char_han(buf, buf2, ch);
@@ -1117,7 +1094,7 @@ void do_score(struct char_data *ch, char *argument, int cmd)
 			strcpy(buf, "You have enough experience to advance.\n\r");
 			strcpy(buf2, "레벨을 올릴만큼 충분한 경험치가 쌓였습니다.\n\r");
 		} else {
-			snprintf(buf, sizeof(buf), "You need %lld experience to advance\n\r",
+			snprintf(buf, sizeof(buf), "You need %lld experience to advance.\n\r",
 				titles[GET_CLASS(ch) - 1][GET_LEVEL(ch) +
 				1].exp -
 				GET_EXP(ch));
@@ -1132,42 +1109,36 @@ void do_score(struct char_data *ch, char *argument, int cmd)
 
 	switch (GET_POS(ch)) {
 	case POSITION_DEAD:
-		send_to_char_han("You are DEAD!\n\r", "당신은 죽으셨습니다\n\r", ch);
+		send_to_char_han("You are DEAD!\n\r", "당신은 죽으셨습니다!\n\r", ch);
 		break;
 	case POSITION_MORTALLYW:
 		send_to_char("You are mortally wounded!, you should seek help!\n\r",
 			     ch);
 		break;
 	case POSITION_INCAP:
-		send_to_char("You are incapacitated, slowly fading away\n\r", ch);
+		send_to_char("You are incapacitated, slowly fading away.\n\r", ch);
 		break;
 	case POSITION_STUNNED:
-		send_to_char("You are stunned! You can't move\n\r", ch);
+		send_to_char("You are stunned! You can't move!\n\r", ch);
 		break;
 	case POSITION_SLEEPING:
-		send_to_char_han("You are sleeping.\n\r",
-				 "당신은 자고 있습니다.\n\r",
-				 ch);
+		send_to_char_han("You are sleeping.\n\r", "당신은 자고 있습니다.\n\r", ch);
 		break;
 	case POSITION_RESTING:
-		send_to_char_han("You are resting.\n\r",
-				 "당신은 쉬고 있습니다.\n\r", ch);
+		send_to_char_han("You are resting.\n\r", "당신은 쉬고 있습니다.\n\r", ch);
 		break;
 	case POSITION_SITTING:
-		send_to_char_han("You are sitting.\n\r",
-				 "당신은 앉아 있습니다.\n\r", ch);
+		send_to_char_han("You are sitting.\n\r", "당신은 앉아 있습니다.\n\r", ch);
 		break;
 	case POSITION_FIGHTING:
 		if (ch->specials.fighting)
-			acthan("You are fighting $N.\n\r",
-			       "당신은 $N님과 싸우고 있습니다.\n\r",
+			acthan("You are fighting $N.\n\r", "당신은 $N님과 싸우고 있습니다.\n\r",
 			       FALSE, ch, 0, ch->specials.fighting, TO_CHAR);
 		else
 			send_to_char("You are fighting thin air.\n\r", ch);
 		break;
 	case POSITION_STANDING:
-		send_to_char_han("You are standing.\n\r",
-				 "당신은 서있습니다.\n\r", ch);
+		send_to_char_han("You are standing.\n\r", "당신은 서있습니다.\n\r", ch);
 		break;
 	default:
 		send_to_char("You are floating.\n\r", ch);
@@ -1175,8 +1146,7 @@ void do_score(struct char_data *ch, char *argument, int cmd)
 	}
 
 	if (ch->affected) {
-		send_to_char_han("Affecting Spells:\n\r",
-				 "걸려있는 마법들:\n\r", ch);
+		send_to_char_han("Affecting Spells:\n\r", "걸려있는 마법들:\n\r", ch);
 		for (aff = ch->affected; aff; aff = aff->next) {
 			snprintf(buf, sizeof(buf), "%s: %d hrs\n\r", spells[aff->type - 1], aff->duration);
 			snprintf(buf2, sizeof(buf2), "%s: %d 시간\n\r", spells[aff->type -
@@ -1185,13 +1155,13 @@ void do_score(struct char_data *ch, char *argument, int cmd)
 		}
 	}
 
-	snprintf(buf, sizeof(buf), "\n\rYou have killed #%d player(s)", ch->player.pk_num);
+	snprintf(buf, sizeof(buf), "\n\rYou have killed #%d player(s).", ch->player.pk_num);
 	snprintf(buf2, sizeof(buf2), "\n\r당신은 지금까지 %d명을 죽였습니다.", ch->player.pk_num);
 	send_to_char_han(buf, buf2, ch);
-	snprintf(buf, sizeof(buf), "\n\rYou have been killed #%d time(s)", ch->player.pked_num);
+	snprintf(buf, sizeof(buf), "\n\rYou have been killed #%d time(s).", ch->player.pked_num);
 	snprintf(buf2, sizeof(buf2), "\n\r당신은 지금까지 %d번 죽었습니다.", ch->player.pked_num);
 	send_to_char_han(buf, buf2, ch);
-	snprintf(buf, sizeof(buf), "\n\rYou have made  #%ld QUEST(s)", ch->quest.solved);
+	snprintf(buf, sizeof(buf), "\n\rYou have made  #%ld QUEST(s).", ch->quest.solved);
 	snprintf(buf2, sizeof(buf2),
 		"\n\r당신은 지금까지 %ld번 QUEST를 풀었습니다.", ch->quest.solved);
 	send_to_char_han(buf, buf2, ch);
@@ -1212,8 +1182,7 @@ void do_attribute(struct char_data *ch, char *argument, int cmd)
 {
 	char buf[MAX_STRING_LENGTH];
 	struct affected_type *aff;
-	extern char *spells[];
-
+	
 	snprintf(buf, sizeof(buf),
 		"You are %d years and %d months, %d cms, and you weigh %d lbs.\n\r",
 		GET_AGE(ch), age(ch).month,
@@ -1276,10 +1245,7 @@ void do_time(struct char_data *ch, char *argument, int cmd)
 {
 	char buf[100], *suf;
 	int weekday, day;
-	extern struct time_info_data time_info;
-	extern char *weekdays[];
-	extern char *month_name[];
-
+	
 	snprintf(buf, sizeof(buf), "It is %d o'clock %s, on ",
 		((time_info.hours % 12 == 0) ? 12 : ((time_info.hours) % 12)),
 		((time_info.hours >= 12) ? "pm" : "am"));
@@ -1320,7 +1286,6 @@ void do_time(struct char_data *ch, char *argument, int cmd)
 
 void do_weather(struct char_data *ch, char *argument, int cmd)
 {
-	extern struct weather_data weather_info;
 	char buf[100], buf2[100];
 	char static *sky_look[4] =
 	{
@@ -1366,11 +1331,6 @@ void do_weather(struct char_data *ch, char *argument, int cmd)
 
 void do_help(struct char_data *ch, char *argument, int cmd)
 {
-	extern int top_of_helpt;
-	extern struct help_index_element *help_index;
-	extern FILE *help_fl;
-	extern char help[MAX_STRING_LENGTH];
-
 	int chk, bot, top, mid, minlen;
 	char buf[MAX_STRING_LENGTH], buffer[MAX_STRING_LENGTH];
 
@@ -1421,9 +1381,7 @@ void do_spells(struct char_data *ch, char *argument, int cmd)
 {
 	char buf[80 * MAX_SKILLS], tmp[MAX_STRING_LENGTH];
 	int no, i;
-	extern char *spells[];
-	extern struct spell_info_type spell_info[];
-
+	
 	if (IS_NPC(ch))
 		return;
 	snprintf(buf, sizeof(buf),
@@ -1469,9 +1427,7 @@ void do_wizhelp(struct char_data *ch, char *argument, int cmd)
 	char buf[MAX_STRING_LENGTH];
 	int no, i, j, is_wizcmd;
 	size_t len = 0;
-	extern char *command[]; 	/* The list of commands (interpreter.c)  */
-	extern struct command_info cmd_info[]; /* cmd_info[1] ~~ commando[0] */
-
+	
 	if (IS_NPC(ch))
 		return;
 	send_to_char("The following privileged commands are available:\n\r", ch);
@@ -1698,7 +1654,6 @@ void do_users(struct char_data *ch, char *argument, int cmd)
     int flag, uptime;
     size_t len = 0;       /* 현재 버퍼에 써진 길이 추적용 */
 	static int most = 0;
-	extern int boottime;
 	
 	one_argument(argument, line);
 	flag = ((GET_LEVEL(ch) < IMO) || (strcmp("-t", line) == 0));
@@ -2008,8 +1963,7 @@ void do_levels(struct char_data *ch, char *argument, int cmd)
 {
 	int i;
 	char buf[MAX_STRING_LENGTH];
-	extern struct title_type titles[4][IMO + 4];
-
+	
 	if (IS_NPC(ch)) {
 		send_to_char("You ain't nothin' but a hound-dog.\n\r", ch);
 		return;
@@ -2127,8 +2081,6 @@ void do_police(struct char_data *ch, char *argument, int cmd)
 
 #ifndef BADDOMS
 #define BADDOMS 16
-extern int baddoms;
-extern char baddomain[BADDOMS][32];
 #endif
 
 /*
