+++ src/guild_command.c
@@ -18,23 +18,7 @@
 #include "limit.h"
 
 #include "guild_list.h"		/* by process */
-/*   external vars  */
-extern struct room_data *world;
-extern struct char_data *character_list;
-extern struct descriptor_data *descriptor_list;
-extern struct index_data *obj_index;
-extern struct time_info_data time_info;
-extern struct title_type titles[4][IMO + 4];
-extern struct index_data *mob_index;
 
-extern char *guild_names[];
-extern int guild_skill_nums[];
-extern char *police_skills[];
-extern char *outlaw_skills[];
-extern char *assasin_skills[];
-extern int police_skill_costs[];
-extern int outlaw_skill_costs[];
-extern int assasin_skill_costs[];
 
 
 /* shuttle bus to KAIST */
