+++ src/magic.c
@@ -20,13 +20,6 @@
 #include "magic_weapon.h"
 #include "prototypes.h"
 
-/* Extern structures */
-extern struct room_data *world;
-extern struct obj_data *object_list;
-extern struct char_data *character_list;
-extern struct index_data *mob_index;
-
-
 
 
 /* Offensive Spells */
@@ -220,8 +213,7 @@ void spell_fireball(byte level, struct char_data *ch,
 		    struct char_data *victim, struct obj_data *obj)
 {
 	int dam;
-	extern struct weather_data weather_info;
-
+	
 	if (!ch || !victim)
 		return;
 
@@ -588,8 +580,6 @@ void spell_call_lightning(byte level, struct char_data *ch,
 {
 	int dam;
 
-	extern struct weather_data weather_info;
-
 	if (!ch || !victim)
 		return;
 
@@ -634,21 +624,9 @@ void spell_harm(byte level, struct char_data *ch,
 void spell_identify(byte level, struct char_data *ch,
 		    struct char_data *victim, struct obj_data *obj)
 {
-	int ac_applicable(struct obj_data *obj_object);
 	char buf[MAX_OUTPUT_LENGTH], buf2[256], bufh[MAX_OUTPUT_LENGTH];
 	int i;
 	bool found;
-
-	struct time_info_data age(struct char_data *ch);
-
-	/* Spell Names */
-	extern char *spells[];
-
-	/* For Objects */
-	extern char *item_types[];
-	extern char *extra_bits[];
-	extern char *apply_types[];
-	extern char *affected_bits[];
 
 	INCREASE_SKILLED2(ch, ch, SPELL_IDENTIFY);
