+++ src/mail.c
@@ -34,7 +34,6 @@ struct mbox_data {
 };
 
 struct mbox_data *mbox_list;
-extern struct room_data *world;
 
 
 
@@ -270,11 +269,7 @@ int post_mail(struct char_data *ch, struct mbox_data *cb, char *arg)
 	char papername[MAX_INPUT_LENGTH], header[MAX_INPUT_LENGTH],
 	buf[MAX_STRING_LENGTH], receiver[MAX_INPUT_LENGTH];
 	struct obj_data *paper;
-	extern struct obj_data *get_obj_in_list_vis(
-							   struct char_data
-							   *ch, char
-							   *name, struct obj_data *list);
-
+	
 	half_chop(arg, receiver, buf);
 	half_chop(buf, papername, header);
 	if (*papername == 0 || *header == 0 || *receiver == 0)
