+++ src/spells1.c
@@ -16,10 +16,6 @@
 
 #include "guild_list.h"
 
-/* Global data */
-extern struct room_data *world;
-extern struct char_data *character_list;
-
 
 
 void cast_burning_hands(byte level, struct char_data *ch, char *arg, int type,
@@ -38,8 +34,6 @@ void cast_burning_hands(byte level, struct char_data *ch, char *arg, int type,
 void cast_call_lightning(byte level, struct char_data *ch, char *arg, int type,
 			 struct char_data *victim, struct obj_data *tar_obj)
 {
-	extern struct weather_data weather_info;
-
 	switch (type) {
 	case SPELL_TYPE_SPELL:
 		if (OUTSIDE(ch) && (weather_info.sky >= SKY_RAINING)) {
