+++ src/spells2.c
@@ -18,12 +18,6 @@
 
 #include "guild_list.h"
 
-/* Global data */
-extern struct room_data *world;
-extern struct char_data *character_list;
-extern struct spell_info_type spell_info[MAX_SPL_LIST];
-extern struct obj_data *object_list;
-
 
 		  
 void cast_armor(byte level, struct char_data *ch, char *arg, int type,
